google-adk/examples/langfuse/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Python template using Restate with Google ADK"
requires-python = ">=3.13,<3.14"

dependencies = [
    "hypercorn",
    "google-adk>=1.18.0",
    "langfuse",
    "openinference-instrumentation-google-adk",
    "restate-sdk",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
]

[tool.uv.sources]
restate-sdk = { path = "../../../../sdk-python/dist/restate_sdk-0.16.0-cp313-cp313-macosx_11_0_arm64.whl" }
